Create new preprocessing function in detect_oscillations and update scripts accordingly.
The new detect_oscillations.preprocessing function's arguments are a SpikingRateMonitor and a PopulationRateMonitor.
It detects whether they are from brian or brian2, then does the preprocessing similarly to the existing
preprocess_spikes function. Calculation without the ISI histogram takes < 1s instead of > 6s.

These changes are not yet included in the automated script. That comes later.